refactor: useDropdown 기능수정
useDropdown 훅 수정. ref를 이용해 dropdownRef 외부의 요소 클릭시에 드롭다운이 닫히도록 변경
