+++ app/compare/_components/compare-page-input.tsx
@@ -4,17 +4,25 @@ import { useEffect, useState } from 'react';
 import CompareTag from '@/components/ui/tags/CompareTag';
 import { SET_PRODUCT } from '@/constants/messages';
 import useCompareStore from '@/store/compareStore';
-import SuggestiveSearchInput, {
-  ProductOption,
-} from '@/app/_styled-guide/_components/suggestive-search-input';
+import { IoMdArrowDropdown } from 'react-icons/io';
+import useSearchSuggestions from '@/hooks/useSearchSuggestions';
+import useDropdown from '@/hooks/useDropdown';
+import DropdownList from '@/app/_styled-guide/_components/dropdown-list';
 
 interface ComparePageInputProps {
   index: number;
 }
 
+export interface ProductOption {
+  id: number;
+  name: string;
+}
+
 function ComparePageInput({ index }: ComparePageInputProps) {
   const [keyword, setKeyword] = useState<string>('');
   const [tag, setTag] = useState<string>('');
+  const { suggestions, isLoading, isError } = useSearchSuggestions(keyword);
+  const { inputRef, dropdownRef, isDropdownOpen, setIsDropdownOpen, handleFocus } = useDropdown();
 
   const { compareItem, updateCompareItem, deleteCompareItem } = useCompareStore((state) => ({
     compareItem: state.compareItems[index],
@@ -29,6 +37,7 @@ function ComparePageInput({ index }: ComparePageInputProps) {
   }, [compareItem, index]);
 
   const onSelect = (option: ProductOption) => {
+    setIsDropdownOpen(false);
     setKeyword('');
     setTag(option.name);
     updateCompareItem(index, { id: option.id, name: option.name });
@@ -39,6 +48,15 @@ function ComparePageInput({ index }: ComparePageInputProps) {
     deleteCompareItem(index);
   };
 
+  // 검색창에서 엔터 입력 시 동작. 검색어는 한 글자 이상 입력해야 함
+  /*
+    const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+      if (e.key === 'Enter' && keyword.trim() && onEnter) {
+        onEnter(keyword);
+      }
+    };
+    */
+
   return (
     <div className="relative flex items-center w-full lg:w-[350px] h-[55px] lg:h-[70px] bg-black-400 rounded-lg">
       {tag && (
@@ -46,14 +64,29 @@ function ComparePageInput({ index }: ComparePageInputProps) {
           <CompareTag productName={tag} onDelete={onTagDelete} index={index} />
         </div>
       )}
-      <SuggestiveSearchInput
-        onSelect={onSelect}
-        keyword={keyword}
-        setKeyword={setKeyword}
-        placeholder={tag ? '' : SET_PRODUCT}
-        inputClassName="w-full bg-transparent text-white"
-        onDisabled={!!tag}
-      />
+
+      <div className="relative flex-1">
+        <div className={`flex justify-between items-center px-5 `}>
+          <input
+            ref={inputRef}
+            type="text"
+            value={keyword}
+            onChange={(e) => setKeyword(e.target.value)}
+            onFocus={handleFocus}
+            placeholder={tag ? '' : SET_PRODUCT}
+            className={`w-full focus:outline-none bg-transparent text-white`}
+            disabled={!!tag}
+          />
+          <button onClick={() => inputRef.current?.focus()}>
+            {!isDropdownOpen && !tag && <IoMdArrowDropdown />}
+          </button>
+        </div>
+        {isDropdownOpen && (
+          <div ref={dropdownRef}>
+            <DropdownList options={suggestions} onSelect={onSelect} className="mt-7" />
+          </div>
+        )}
+      </div>
     </div>
   );
 }
